+++ pcsx2/CDVD/CDVD.cpp
@@ -264,7 +264,7 @@ void cdvdSaveNVRAM(void)
 static void cdvdReadNVM(u8* dst, int offset, int bytes)
 {
 	int to_read = bytes;
-	if ((offset + bytes) > sizeof(s_nvram))
+	if ((offset + bytes) > static_cast<int>(sizeof(s_nvram)))
 	{
 		Console.Warning("CDVD: Out of bounds NVRAM read: offset=%d, bytes=%d", offset, bytes);
 		to_read = std::max(static_cast<int>(sizeof(s_nvram)) - offset, 0);
@@ -278,7 +278,7 @@ static void cdvdReadNVM(u8* dst, int offset, int bytes)
 static void cdvdWriteNVM(const u8* src, int offset, int bytes)
 {
 	int to_write = bytes;
-	if ((offset + bytes) > sizeof(s_nvram))
+	if ((offset + bytes) > static_cast<int>(sizeof(s_nvram)))
 	{
 		Console.Warning("CDVD: Out of bounds NVRAM write: offset=%d, bytes=%d", offset, bytes);
 		to_write = std::max(static_cast<int>(sizeof(s_nvram)) - offset, 0);
